+++ images/tools/bazel-dind-4.2-ubuntu/tests/run-tests.sh
@@ -6,3 +6,9 @@ echo "[INFO] Run tests"
 
 echo '[TEST] Print out bazel version' && \
     bazel --version
+
+echo '[TEST] Check dockerd version' && \
+    dockerd --version
+
+echo '[TEST] Check docker version' && \
+    docker --version