+++ images/tools/bazel-dind-4.2-ubuntu/CHANGELOG.md
@@ -1,4 +1,4 @@
 # CHANGELOG
 
 - bazel 4.2.1
-- docker
+- docker:stable ()